requirements.txt
Django==4.1.3
django-rest-fuzzysearch==0.5.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
Pillow==9.3.0
psycopg2==2.9.5
whitenoise